Separate UDP and Quic TPU connection implementations
